Moved error parsing logic to plugin
