+++ src/main/java/com/exceptionless/exceptionlessclient/plugins/preconfigured/args/RequestInfoGetArgs.java
@@ -1,4 +1,4 @@
-package com.exceptionless.exceptionlessclient.services;
+package com.exceptionless.exceptionlessclient.plugins.preconfigured.args;
 
 import lombok.Builder;
 import lombok.Value;
