--- src/main/java/com/exceptionless/exceptionlessclient/services/ErrorParserIF.java
@@ -1,7 +1,0 @@
-package com.exceptionless.exceptionlessclient.services;
-
-import com.exceptionless.exceptionlessclient.models.services.error.Error;
-
-public interface ErrorParserIF {
-    Error parse(Exception exception);
-}
